+++ international_online_offer/views.py
@@ -3,3 +3,7 @@ from django.views.generic import TemplateView
 
 class IOOIndex(TemplateView):
     template_name = 'ioo/index.html'
+
+
+class IOOSector(TemplateView):
+    template_name = 'ioo/triage/sector.html'